Avoid an infinity loop in LineObserver()
Currently if somehow terminator bytes haven't been found
in the buffer, the function just continues the outer loop,
but it can't acquire more because the buffer isn't empty.
So we must first try to acquire bytes and if we get nothing
then we will just replace the buffer with an empty byte array.
Or else we will have a chance to find terminator bytes this time